gro: smart search refactor, improve popup
