+++ extension/base/core/date_utils.mjs
@@ -662,11 +662,32 @@ const DateUtils = {
 
     let parts = yearMonthDayString.split(separator);
 
-    if (parts.length != 3 || parts[0].length != 4) {
+    if (parts.length == 0 || parts[0].length != 4) {
       return "";
     }
 
-    return this.getDateStringFromYearMonthDay(parts[0], parts[1], parts[2]);
+    let year = Number(parts[0]);
+    if (isNaN(year)) {
+      year = 0;
+    }
+
+    let month = 0;
+    if (parts.length > 1) {
+      month = Number(parts[1]);
+      if (isNaN(month)) {
+        month = 0;
+      }
+    }
+
+    let day = 0;
+    if (parts.length > 2) {
+      day = Number(parts[2]);
+      if (isNaN(day)) {
+        day = 0;
+      }
+    }
+
+    return this.getDateStringFromYearMonthDay(year, month, day);
   },
 
   compareParsedDates: function (parsedDateA, parsedDateB) {
